+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/addons/Filtering/FilterFlyout.js
@@ -2,45 +2,20 @@
 /*
  * Licensed Materials - Property of IBM
  * 5724-Q36
- * (c) Copyright IBM Corp. 2022
+ * (c) Copyright IBM Corp. 2023
  * US Government Users Restricted Rights - Use, duplication or disclosure
  * restricted by GSA ADP Schedule Contract with IBM Corp.
  */
 import { Filter } from '@carbon/react/icons';
-import {
-  Button,
-  Checkbox,
-  DatePicker,
-  DatePickerInput,
-  Dropdown,
-  FormGroup,
-  NumberInput,
-  RadioButton,
-  RadioButtonGroup,
-  usePrefix,
-} from '@carbon/react';
+import { Button, usePrefix } from '@carbon/react';
 import cx from 'classnames';
 import PropTypes from 'prop-types';
-import React, { useCallback, useRef, useState } from 'react';
+import React, { useRef, useState } from 'react';
 import { useClickOutside } from '../../../../../global/js/hooks';
 import { pkg } from '../../../../../settings';
 import { ActionSet } from '../../../../ActionSet';
-import {
-  BATCH,
-  CHECKBOX,
-  CLEAR_FILTERS,
-  DATE,
-  DROPDOWN,
-  FLYOUT,
-  INSTANT,
-  NUMBER,
-  RADIO,
-} from './constants';
-import {
-  useInitialStateFromFilters,
-  useSubscribeToEventEmitter,
-} from './hooks';
-import { getInitialStateFromFilters } from './utils';
+import { BATCH, CLEAR_FILTERS, FLYOUT, INSTANT } from './constants';
+import { useSubscribeToEventEmitter, useFilters } from './hooks';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 const componentClass = `${blockClass}-filter-flyout`;
@@ -61,17 +36,24 @@ const FilterFlyout = ({
 }) => {
   /** State */
   const [open, setOpen] = useState(false);
-  const [filtersState, setFiltersState] = useInitialStateFromFilters(
+
+  const {
+    filtersState,
+    prevFiltersObjectArrayRef,
+    prevFiltersRef,
+    revertToPreviousFilters,
+    reset,
+    renderFilter,
+    filtersObjectArray,
+  } = useFilters({
+    updateMethod,
     filters,
-    FLYOUT
-  );
-  const [filtersObjectArray, setFiltersObjectArray] = useState([]);
+    setAllFilters,
+    variation: FLYOUT,
+  });
 
   /** Refs */
   const filterFlyoutRef = useRef(null);
-  // When using batch actions we have to store the filters to then apply them later
-  const prevFiltersRef = useRef(JSON.stringify(filtersState));
-  const prevFiltersObjectArrayRef = useRef(JSON.stringify(filtersObjectArray));
 
   /** Memos */
   const showActionSet = updateMethod === BATCH;
@@ -107,77 +89,6 @@ const FilterFlyout = ({
     closeFlyout();
   };
 
-  const reset = () => {
-    // Get the initial values for the filters
-    const initialFiltersState = getInitialStateFromFilters(filters, FLYOUT);
-    const initialFiltersObjectArray = [];
-
-    // Set the state to the initial values
-    setFiltersState(initialFiltersState);
-    setFiltersObjectArray(initialFiltersObjectArray);
-    setAllFilters([]);
-
-    // Update their respective refs so everything is in sync
-    prevFiltersRef.current = JSON.stringify(initialFiltersState);
-    prevFiltersObjectArrayRef.current = JSON.stringify(
-      initialFiltersObjectArray
-    );
-  };
-
-  // If the user decides to cancel or click outside the flyout, it reverts back to the filters that were
-  // there when they opened the flyout
-  const revertToPreviousFilters = () => {
-    setFiltersState(JSON.parse(prevFiltersRef.current));
-    setFiltersObjectArray(JSON.parse(prevFiltersObjectArrayRef.current));
-  };
-
-  const applyFilters = useCallback(
-    ({ column, value, type }) => {
-      // If no end date is selected return because we need the end date to do computations
-      if (type === DATE && !value[1]) {
-        return;
-      }
-
-      const filtersObjectArrayCopy = [...filtersObjectArray];
-      // check if the filter already exists in the array
-      const filter = filtersObjectArrayCopy.find((item) => item.id === column);
-
-      // if filter exists in array then update the filter's new value
-      if (filter) {
-        filter.value = value;
-      } else {
-        filtersObjectArrayCopy.push({ id: column, value, type });
-      }
-
-      if (type === CHECKBOX) {
-        /**
-        When all checkboxes of a group are all unselected the value still exists in the filtersObjectArray
-        This checks if all the checkboxes are selected = false and removes it from the array
-       */
-        const index = filtersObjectArrayCopy.findIndex(
-          (filter) => filter.id === column
-        );
-
-        // If all the selected state is false remove from array
-        const shouldRemoveFromArray = filtersObjectArrayCopy[index].value.every(
-          (val) => val.selected === false
-        );
-
-        if (shouldRemoveFromArray) {
-          filtersObjectArrayCopy.splice(index, 1);
-        }
-      }
-
-      setFiltersObjectArray(filtersObjectArrayCopy);
-
-      // Automatically apply the filters if the updateMethod is instant
-      if (updateMethod === INSTANT) {
-        setAllFilters(filtersObjectArrayCopy);
-      }
-    },
-    [setAllFilters, updateMethod, filtersObjectArray]
-  );
-
   /** Effects */
   useClickOutside(filterFlyoutRef, (target) => {
     const hasClickedOnDatePicker = target.closest('.flatpickr-calendar');
@@ -192,141 +103,6 @@ const FilterFlyout = ({
   });
 
   useSubscribeToEventEmitter(CLEAR_FILTERS, reset);
-
-  /** Render the individual filter component */
-  const renderFilter = ({ type, column, props: components }) => {
-    if (type === DATE) {
-      return (
-        <DatePicker
-          {...components.DatePicker}
-          onChange={(value) => {
-            setFiltersState({ ...filtersState, [column]: { value, type } });
-            applyFilters({ column, value, type });
-            components.DatePicker.onChange?.(value);
-          }}
-          value={filtersState[column].value}
-          datePickerType="range"
-        >
-          <DatePickerInput
-            placeholder="mm/dd/yyyy"
-            labelText="Start date"
-            {...components.DatePickerInput.start}
-          />
-          <DatePickerInput
-            placeholder="mm/dd/yyyy"
-            labelText="End date"
-            {...components.DatePickerInput.end}
-          />
-        </DatePicker>
-      );
-    } else if (type === NUMBER) {
-      return (
-        <NumberInput
-          step={1}
-          allowEmpty
-          hideSteppers
-          {...components.NumberInput}
-          onChange={(event) => {
-            setFiltersState({
-              ...filtersState,
-              [column]: {
-                value: event.target.value,
-                type,
-              },
-            });
-            applyFilters({ column, value: event.target.value, type });
-            components.NumberInput.onChange?.(event);
-          }}
-          value={filtersState[column].value}
-        />
-      );
-    } else if (type === CHECKBOX) {
-      return (
-        <FormGroup {...components.FormGroup}>
-          {filtersState[column].value.map((option) => (
-            <Checkbox
-              key={option.labelText}
-              {...option}
-              onChange={(_, { checked: isSelected }) => {
-                const checkboxCopy = filtersState[column].value;
-                const foundCheckbox = checkboxCopy.find(
-                  (checkbox) => checkbox.value === option.value
-                );
-                foundCheckbox.selected = isSelected;
-                setFiltersState({
-                  ...filtersState,
-                  [column]: {
-                    value: checkboxCopy,
-                    type,
-                  },
-                });
-                applyFilters({
-                  column,
-                  value: [...filtersState[column].value],
-                  type,
-                });
-                option.onChange?.(isSelected);
-              }}
-              checked={option.selected}
-            />
-          ))}
-        </FormGroup>
-      );
-    } else if (type === RADIO) {
-      return (
-        <FormGroup {...components.FormGroup}>
-          <RadioButtonGroup
-            {...components.RadioButtonGroup}
-            valueSelected={filtersState[column].value}
-            onChange={(selected) => {
-              setFiltersState({
-                ...filtersState,
-                [column]: {
-                  value: selected,
-                  type,
-                },
-              });
-              applyFilters({
-                column,
-                value: selected,
-                type,
-              });
-              components.RadioButtonGroup.onChange?.(selected);
-            }}
-          >
-            {components.RadioButton.map((radio) => (
-              <RadioButton
-                key={radio.id ?? radio.labelText ?? radio.value}
-                {...radio}
-              />
-            ))}
-          </RadioButtonGroup>
-        </FormGroup>
-      );
-    } else if (type === DROPDOWN) {
-      return (
-        <Dropdown
-          {...components.Dropdown}
-          selectedItem={filtersState[column].value}
-          onChange={({ selectedItem }) => {
-            setFiltersState({
-              ...filtersState,
-              [column]: {
-                value: selectedItem,
-                type,
-              },
-            });
-            applyFilters({
-              column,
-              value: selectedItem,
-              type,
-            });
-            components.Dropdown.onChange?.(selectedItem);
-          }}
-        />
-      );
-    }
-  };
 
   /** Renders all filters */
   const renderFilters = () => filters.map(renderFilter);
