+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/addons/Filtering/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useInitialStateFromFilters } from './useInitialStateFromFilters';
 export { default as useSubscribeToEventEmitter } from './useSubscribeToEventEmitter';
+export { default as useFilters } from './useFilters';
