+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/addons/Filtering/utils.js
@@ -5,14 +5,22 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { FLYOUT, PANEL } from './constants';
+import {
+  CHECKBOX,
+  DATE,
+  DROPDOWN,
+  FLYOUT,
+  NUMBER,
+  PANEL,
+  RADIO,
+} from './constants';
 
 // This functions takes the filters passed in and makes an object to track it's state
 export const getInitialStateFromFilters = (filters, variation = FLYOUT) => {
   const initialFilterState = {};
 
   const setInitialState = ({ type, column, props }) => {
-    if (type === 'checkbox') {
+    if (type === CHECKBOX) {
       initialFilterState[column] = {
         value: props.Checkbox.map(({ id, labelText, value }) => ({
           id,
@@ -22,22 +30,22 @@ export const getInitialStateFromFilters = (filters, variation = FLYOUT) => {
         })),
         type,
       };
-    } else if (type === 'date') {
+    } else if (type === DATE) {
       initialFilterState[column] = {
         value: [null, null],
         type,
       };
-    } else if (type === 'number') {
+    } else if (type === NUMBER) {
       initialFilterState[column] = {
         value: '',
         type,
       };
-    } else if (type === 'radio') {
+    } else if (type === RADIO) {
       initialFilterState[column] = {
         value: '',
         type,
       };
-    } else if (type === 'dropdown') {
+    } else if (type === DROPDOWN) {
       initialFilterState[column] = {
         value: '',
         type,
